fixed: bad type cast

+++ xbmc/osx/WinSystemOSX.mm
@@ -37,7 +37,6 @@
 #import <QuartzCore/QuartzCore.h>
 #import <Carbon/Carbon.h>   // ShowMenuBar, HideMenuBar
 
-
 #define MAX_DISPLAYS 32
 static NSWindow* blankingWindows[MAX_DISPLAYS];
 
@@ -737,7 +736,7 @@ void CWinSystemOSX::EnableVSync(bool enable)
   GLint swapInterval;
   
   swapInterval = enable ? 1 : 0;
-  [[NSOpenGLContext currentContext] setValues:(const long*)&swapInterval forParameter:NSOpenGLCPSwapInterval];
+  [[NSOpenGLContext currentContext] setValues:(const GLint*)&swapInterval forParameter:NSOpenGLCPSwapInterval];
 }
 
 bool CWinSystemOSX::SwitchToVideoMode(int width, int height, double refreshrate)
